- Moved automapper into the app and change the import paths
