+++ app/__init__.py
@@ -1,4 +1,4 @@
-from automapper import create_mapper
+from app.automapper import create_mapper
 
 # create another mapper for viewmodels
 viewmodel_mapper = create_mapper()
